core/standards/css3/image-fit/js/getComputedStyle/preserveAspectRatio/xMidYMin_meet.html

<!DOCTYPE html>
<html>
	<head>
		<title>getComputedStyle SVG preserveAspectratio='xMidYMin meet'</title>
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
	</head>
	<body>
		<img id="test" src="xMidYMin_meet.svg">
		<script>
		test(function() {
		  var e = document.getElementById('test');
		  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
		  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
		});
		</script>
	</body>
</html>
